% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfile_utils.R
\name{add_patch_areas}
\alias{add_patch_areas}
\title{Add patch areas from worldfile}
\usage{
add_patch_areas(output_df, world_source)
}
\arguments{
\item{output_df}{Data table with patchID column}

\item{world_source}{Path to worldfile or data frame}
}
\description{
Add patch areas from worldfile
}
