% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfile_utils.R
\name{add_patch_veg_parm_IDs}
\alias{add_patch_veg_parm_IDs}
\title{Add patch vegetation parameter IDs from worldfile}
\usage{
add_patch_veg_parm_IDs(output_df, world_source)
}
\arguments{
\item{output_df}{Data table with patchID column}

\item{world_source}{Path to worldfile or data frame}
}
\description{
Add patch vegetation parameter IDs from worldfile
}
