% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfile_utils.R
\name{add_patchfamily_IDarea}
\alias{add_patchfamily_IDarea}
\title{Add patch family IDs and areas}
\usage{
add_patchfamily_IDarea(output_df)
}
\arguments{
\item{output_df}{Data table with patchID, area, year, month}
}
\description{
Add patch family IDs and areas
}
