% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tec_utils.R
\name{add_tec}
\alias{add_tec}
\title{Add a tec event}
\usage{
add_tec(input_tec_data, date, tecname)
}
\arguments{
\item{input_tec_data}{Tec data frame}

\item{date}{Date-time string in RHESSys format}

\item{tecname}{Event name}
}
\description{
Add a tec event
}
