% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AddRunIDtoOutputFilters}
\alias{AddRunIDtoOutputFilters}
\title{Append run ID to output filters}
\usage{
AddRunIDtoOutputFilters(output_filter, runid)
}
\arguments{
\item{output_filter}{List of output filters}

\item{runid}{Run identifier string/number}
}
\description{
Append run ID to output filters
}
