% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{agg_dyn}
\alias{agg_dyn}
\title{Aggregate dynamically by variables}
\usage{
agg_dyn(DT, name, aggvars, vars)
}
\arguments{
\item{DT}{Data table to aggregate}

\item{name}{Run identifier}

\item{aggvars}{Grouping variables}

\item{vars}{Variables to aggregate}
}
\description{
Aggregate dynamically by variables
}
