% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{agg_YM_strata}
\alias{agg_YM_strata}
\title{Aggregate by year-month and strata}
\usage{
agg_YM_strata(X, Y)
}
\arguments{
\item{X}{Data table to aggregate}

\item{Y}{Run identifier}
}
\description{
Aggregate by year-month and strata
}
