% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{basin2utm}
\alias{basin2utm}
\title{Determine UTM zone for basin raster}
\usage{
basin2utm(basin)
}
\arguments{
\item{basin}{SpatRaster basin mask}
}
\description{
Determine UTM zone for basin raster
}
