% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{basin_daily2mn}
\alias{basin_daily2mn}
\title{Aggregate basin daily outputs to monthly}
\usage{
basin_daily2mn(basin_daily)
}
\arguments{
\item{basin_daily}{Data table of daily basin outputs}
}
\description{
Aggregate basin daily outputs to monthly
}
