% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{basin_output_fixes}
\alias{basin_output_fixes}
\title{Apply fixes to basin outputs}
\usage{
basin_output_fixes(DT)
}
\arguments{
\item{DT}{Data table of basin outputs}
}
\description{
Apply fixes to basin outputs
}
