% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_redefine2.R
\name{build_redefine2}
\alias{build_redefine2}
\title{build_redefine2}
\usage{
build_redefine2(
  worldfile,
  out_file = NULL,
  vars = NULL,
  values = NULL,
  std_thin = NULL,
  patchID = NULL,
  strataID = NULL,
  veg_parm_ID = NULL
)
}
\description{
build_redefine2
}
