% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_redefine3.R
\name{build_redefine3}
\alias{build_redefine3}
\title{build_redefine3}
\usage{
build_redefine3(
  worldfile,
  out_file = NULL,
  vars = NULL,
  values = NULL,
  std_thin = NULL,
  patchID = NULL,
  strataID = NULL,
  veg_parm_ID = NULL,
  add = FALSE
)
}
\description{
One more try better now
}
