% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_defpar_sens.R
\name{cal_defpar_sens}
\alias{cal_defpar_sens}
\title{cal_defpar_sens}
\usage{
cal_defpar_sens(
  out_dir,
  obs_source,
  input_def_pars = NULL,
  monthly = F,
  sortby = "NSE",
  add_base = T,
  def_changes_by_stat = F
)
}
\description{
Sensetivity of of calibration evaluation statistics (NSE, NSElog, PBIAS, RMSE, r2) to changes in defintion file parameters.
}
