% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{check_template}
\alias{check_template}
\title{Summarize RHESSys template contents}
\usage{
check_template(template, map_dir = NULL)
}
\arguments{
\item{template}{Path to RHESSys template file}

\item{map_dir}{Optional directory containing input map rasters to check they exist}
}
\description{
Summarize RHESSys template contents
}
