% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_utils.R
\name{cleanup_wmfire}
\alias{cleanup_wmfire}
\title{Move WMFire grids into timestamped folder}
\usage{
cleanup_wmfire(dir)
}
\arguments{
\item{dir}{Directory containing WMFire grid outputs}
}
\description{
Move WMFire grids into timestamped folder
}
