% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_utils.R
\name{clim_repeat}
\alias{clim_repeat}
\title{Repeat a climate file n times}
\usage{
clim_repeat(clim, file, n, units = "years")
}
\arguments{
\item{clim}{Path to climate file or data.frame}

\item{file}{Output file prefix}

\item{n}{Number of repetitions}

\item{units}{Units for repetition (e.g., "years")}
}
\description{
Repeat a climate file n times
}
