% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{collapse_rh_input_files_list}
\alias{collapse_rh_input_files_list}
\title{Collapse nested RHESSys input file lists}
\usage{
collapse_rh_input_files_list(filelist)
}
\arguments{
\item{filelist}{Nested list of RHESSys input files}
}
\description{
Collapse nested RHESSys input file lists
}
