% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_utils.R
\name{collect_csvs}
\alias{collect_csvs}
\title{Collect CSV outputs into a new folder}
\usage{
collect_csvs(dir, dir_base = "rh_out_")
}
\arguments{
\item{dir}{Directory containing CSV outputs}

\item{dir_base}{Base name for output directory}
}
\description{
Collect CSV outputs into a new folder
}
