% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_utils.R
\name{collect_params}
\alias{collect_params}
\title{Collect params files into a new folder}
\usage{
collect_params(dir)
}
\arguments{
\item{dir}{Directory containing *.params files}
}
\description{
Collect params files into a new folder
}
