% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{copy_param_by_var}
\alias{copy_param_by_var}
\title{Copy parameter values by variable}
\usage{
copy_param_by_var(pars, copy_var, replace_var)
}
\arguments{
\item{pars}{Parameter list}

\item{copy_var}{Variable name to copy from}

\item{replace_var}{Variable name to replace}
}
\description{
Copy parameter values by variable
}
