% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{daily_dates}
\alias{daily_dates}
\title{Add run and year-month indices}
\usage{
daily_dates(X, Y)
}
\arguments{
\item{X}{Data frame with year/month columns}

\item{Y}{Run identifier}
}
\description{
Add run and year-month indices
}
