% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_utils.R
\name{def_changes_by_evaldf}
\alias{def_changes_by_evaldf}
\title{Combine definition changes with evaluation statistics}
\usage{
def_changes_by_evaldf(defpar_df, eval, stat = "NSE")
}
\arguments{
\item{defpar_df}{Data frame of definition parameter changes}

\item{eval}{Data frame of evaluation metrics}

\item{stat}{Statistic name used for ordering}
}
\description{
Combine definition changes with evaluation statistics
}
