% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{def_par_allcomb}
\alias{def_par_allcomb}
\title{Generate all parameter combinations}
\usage{
def_par_allcomb(defpars)
}
\arguments{
\item{defpars}{List of parameter options}
}
\description{
Generate all parameter combinations
}
