% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{defpar_df_t2parcols}
\alias{defpar_df_t2parcols}
\title{Pivot definition parameter data frame to columns}
\usage{
defpar_df_t2parcols(defpar_df)
}
\arguments{
\item{defpar_df}{Definition parameter data frame}
}
\description{
Pivot definition parameter data frame to columns
}
