% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{defpar_extract_byrunnum}
\alias{defpar_extract_byrunnum}
\title{Extract parameters by run number}
\usage{
defpar_extract_byrunnum(pars_list, runnum)
}
\arguments{
\item{pars_list}{List of definition parameters}

\item{runnum}{Run identifier}
}
\description{
Extract parameters by run number
}
