% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{defpar_rm_dup}
\alias{defpar_rm_dup}
\title{Remove duplicate definition parameters}
\usage{
defpar_rm_dup(def_list)
}
\arguments{
\item{def_list}{List of definition parameters}
}
\description{
Remove duplicate definition parameters
}
