% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{defpars_csv2list}
\alias{defpars_csv2list}
\title{Convert parameter CSV to list}
\usage{
defpars_csv2list(defpar_csv)
}
\arguments{
\item{defpar_csv}{Path to parameter CSV file}
}
\description{
Convert parameter CSV to list
}
