% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{defpars_list2df}
\alias{defpars_list2df}
\title{Convert defpar list to data frame}
\usage{
defpars_list2df(defpars)
}
\arguments{
\item{defpars}{List of definition parameters}
}
\description{
Convert defpar list to data frame
}
