% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{dup_soil_pars}
\alias{dup_soil_pars}
\title{Duplicate soil parameters}
\usage{
dup_soil_pars(input_def_pars, input_hdr)
}
\arguments{
\item{input_def_pars}{Definition parameter list}

\item{input_hdr}{Header lines to prepend}
}
\description{
Duplicate soil parameters
}
