% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{extract_world}
\alias{extract_world}
\title{Extract a sub-world by unique ID}
\usage{
extract_world(world, target_unique_ID)
}
\arguments{
\item{world}{Worldfile data frame}

\item{target_unique_ID}{Unique ID to extract around}
}
\description{
Extract a sub-world by unique ID
}
