% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_open_most_recent}
\alias{find_open_most_recent}
\title{Open the most recent file matching a pattern}
\usage{
find_open_most_recent(dir, pattern)
}
\arguments{
\item{dir}{Directory to search}

\item{pattern}{Filename pattern (regex)}
}
\description{
Open the most recent file matching a pattern
}
