% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{g_legend}
\alias{g_legend}
\title{Extract a legend grob from a ggplot object}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot object}
}
\value{
A gtable containing the plot legend
}
\description{
Extract a legend grob from a ggplot object
}
