% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_basin_daily}
\alias{get_basin_daily}
\title{Load basin daily outputs}
\usage{
get_basin_daily(out_dir)
}
\arguments{
\item{out_dir}{Directory containing basin CSV outputs}
}
\description{
Load basin daily outputs
}
