% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{get_def_par}
\alias{get_def_par}
\title{Get a parameter from a definition file}
\usage{
get_def_par(def_file, parameter = NULL)
}
\arguments{
\item{def_file}{Path to definition file}

\item{parameter}{Optional parameter name to extract}
}
\description{
Get a parameter from a definition file
}
