% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{get_only_varied_defpars_list}
\alias{get_only_varied_defpars_list}
\title{Get only varied definition parameters}
\usage{
get_only_varied_defpars_list(defpars)
}
\arguments{
\item{defpars}{List of definition parameters}
}
\description{
Get only varied definition parameters
}
