% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_patch_daily}
\alias{get_patch_daily}
\title{Load patch daily outputs}
\usage{
get_patch_daily(out_dir)
}
\arguments{
\item{out_dir}{Directory containing patch CSV outputs}
}
\description{
Load patch daily outputs
}
