% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_stratum_daily}
\alias{get_stratum_daily}
\title{Load stratum daily outputs}
\usage{
get_stratum_daily(out_dir)
}
\arguments{
\item{out_dir}{Directory containing stratum CSV outputs}
}
\description{
Load stratum daily outputs
}
