% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetRunID}
\alias{GetRunID}
\title{Get or update run ID}
\usage{
GetRunID(increment = FALSE, reset = FALSE)
}
\arguments{
\item{increment}{Logical; increment stored run ID}

\item{reset}{Logical; reset stored run ID to zero}
}
\description{
Get or update run ID
}
