% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{input_map_plotpdf}
\alias{input_map_plotpdf}
\title{Plot preprocessing input maps to a PDF}
\usage{
input_map_plotpdf(
  map_dir,
  template = NULL,
  out_name = "map_plots",
  dest = "preprocessing",
  map_exts = c(".tif|.tiff"),
  pdfwidth = 7,
  pdfheight = 7
)
}
\arguments{
\item{map_dir}{Directory containing input map rasters}

\item{template}{Optional RHESSys template path for map metadata}

\item{out_name}{Base filename for the PDF}

\item{dest}{Output directory for the PDF}

\item{map_exts}{Regex pattern of raster filename extensions}

\item{pdfwidth}{PDF width in inches}

\item{pdfheight}{PDF height in inches}
}
\description{
Plot preprocessing input maps to a PDF
}
