% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{list_rh_input_files}
\alias{list_rh_input_files}
\title{List RHESSys input files required for a run}
\usage{
list_rh_input_files(
  input_rhessys,
  input_hdr,
  input_def_pars,
  input_tec_data,
  output_filter
)
}
\arguments{
\item{input_rhessys}{RHESSys inputs list}

\item{input_hdr}{Header list of definition files}

\item{input_def_pars}{Definition parameter list}

\item{input_tec_data}{Tec events data frame}

\item{output_filter}{Output filter list}
}
\description{
List RHESSys input files required for a run
}
