% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{LPC2std_veg_cover}
\alias{LPC2std_veg_cover}
\title{Convert LPC to standard vegetation cover IDs}
\usage{
LPC2std_veg_cover(
  lpc_df,
  mask_map_path,
  lpc_map_list,
  vegid_tree = 1,
  vegid_shrub = 5,
  vegid_herb = 3,
  vegid_other = 4
)
}
\arguments{
\item{lpc_df}{LPC data frame with cover percentages}

\item{mask_map_path}{Path to mask raster}

\item{lpc_map_list}{List of LPC rasters}

\item{vegid_tree}{Veg ID for tree class}

\item{vegid_shrub}{Veg ID for shrub class}

\item{vegid_herb}{Veg ID for herb class}

\item{vegid_other}{Veg ID for other class}
}
\description{
Convert LPC to standard vegetation cover IDs
}
