% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{lpc_as_df}
\alias{lpc_as_df}
\title{Convert LPC rasters to data frame}
\usage{
lpc_as_df(lpc_map_list, plots = F)
}
\arguments{
\item{lpc_map_list}{List of LPC raster layers}

\item{plots}{Logical; show histograms}
}
\description{
Convert LPC rasters to data frame
}
