% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{lpc_bin}
\alias{lpc_bin}
\title{Bin LPC percentages}
\usage{
lpc_bin(lpc_df, round_value)
}
\arguments{
\item{lpc_df}{LPC data frame}

\item{round_value}{Rounding increment}
}
\description{
Bin LPC percentages
}
