% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{lpc_extract}
\alias{lpc_extract}
\title{Extract and prep Landfire/LPC tiles}
\usage{
lpc_extract(lpc_data_path, lpc_pattern, mask_map_path, plots = F)
}
\arguments{
\item{lpc_data_path}{Directory containing LPC tiles}

\item{lpc_pattern}{Vector of tile patterns}

\item{mask_map_path}{Path to mask raster}

\item{plots}{Logical; show diagnostic plots}
}
\description{
Extract and prep Landfire/LPC tiles
}
