% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{make_1pfamflow}
\alias{make_1pfamflow}
\title{Create a single patch-family flowtable string}
\usage{
make_1pfamflow(world)
}
\arguments{
\item{world}{Worldfile data frame}
}
\description{
Create a single patch-family flowtable string
}
