% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{make_1pflow}
\alias{make_1pflow}
\title{Create a single-patch flowtable string}
\usage{
make_1pflow(world)
}
\arguments{
\item{world}{Worldfile data frame}
}
\description{
Create a single-patch flowtable string
}
