% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{make_par_list_from_def_file}
\alias{make_par_list_from_def_file}
\title{Make parameter list from definition file}
\usage{
make_par_list_from_def_file(def_file, parameters, defaults = NULL)
}
\arguments{
\item{def_file}{Path to definition file}

\item{parameters}{Parameter names to extract}

\item{defaults}{Optional default values}
}
\description{
Make parameter list from definition file
}
