% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{make_rules_map}
\alias{make_rules_map}
\title{Create rules map raster}
\usage{
make_rules_map(lpc_rules_df, mask_map_path, lpc_map_list)
}
\arguments{
\item{lpc_rules_df}{Rules data frame}

\item{mask_map_path}{Path to mask raster}

\item{lpc_map_list}{List of LPC rasters}
}
\description{
Create rules map raster
}
