% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{make_std_rules}
\alias{make_std_rules}
\title{Make standard rule table}
\usage{
make_std_rules(lpc_bin_df)
}
\arguments{
\item{lpc_bin_df}{Binned LPC data frame}
}
\description{
Make standard rule table
}
