% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{make_thin_rule}
\alias{make_thin_rule}
\title{Create thinned vegetation rule table}
\usage{
make_thin_rule(unique_rules, tree_thin_pct = 0, shrub_thin_pct = 0)
}
\arguments{
\item{unique_rules}{Data frame of unique rules}

\item{tree_thin_pct}{Proportion of tree cover to thin}

\item{shrub_thin_pct}{Proportion of shrub cover to thin}
}
\description{
Create thinned vegetation rule table
}
