% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{new_world_from_spun_worlds}
\alias{new_world_from_spun_worlds}
\title{Combine spun worlds into a new worldfile}
\usage{
new_world_from_spun_worlds(
  spun_world_paths,
  original_world_path,
  ID = "rule_ID"
)
}
\arguments{
\item{spun_world_paths}{Vector of spun worldfile paths}

\item{original_world_path}{Path to the original worldfile}

\item{ID}{Identifier column to align on (`"rule_ID"` or `"veg_parm_ID"`)}
}
\description{
Combine spun worlds into a new worldfile
}
