% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_utils.R
\name{par_ranges_by_pctle}
\alias{par_ranges_by_pctle}
\title{Compute parameter ranges for top-performing runs}
\usage{
par_ranges_by_pctle(defchg_nse, pct = 0.9)
}
\arguments{
\item{defchg_nse}{Combined parameter/statistics table}

\item{pct}{Quantile threshold used to select runs}
}
\description{
Compute parameter ranges for top-performing runs
}
