% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_utils.R
\name{parallel_runs}
\alias{parallel_runs}
\title{Execute RHESSys commands in parallel}
\usage{
parallel_runs(run_cmds)
}
\arguments{
\item{run_cmds}{List of command strings to execute}
}
\description{
Execute RHESSys commands in parallel
}
