% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{pars_sens}
\alias{pars_sens}
\title{Sensitivity of parameters}
\usage{
pars_sens(out_dir, input_def_pars, sortby = "Mean", omit_mising_data = T)
}
\arguments{
\item{out_dir}{Output directory with simulation results}

\item{input_def_pars}{Definition parameter list}

\item{sortby}{Statistic to sort by}

\item{omit_mising_data}{Logical; omit runs with missing data}
}
\description{
Sensitivity of parameters
}
