% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_utils.R
\name{pars_sens_output_tables}
\alias{pars_sens_output_tables}
\title{Save parameter sensitivity tables to a PDF}
\usage{
pars_sens_output_tables(
  pars_sens_out,
  output_path = "pars_sens_tables.pdf",
  pdfwidth = 14,
  pdfheight = 14
)
}
\arguments{
\item{pars_sens_out}{List of sensitivity result tables}

\item{output_path}{Output PDF path}

\item{pdfwidth}{PDF width in inches}

\item{pdfheight}{PDF height in inches}
}
\description{
Save parameter sensitivity tables to a PDF
}
