% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_utils.R
\name{parse_rhessys_cmdout}
\alias{parse_rhessys_cmdout}
\title{Parse RHESSys command output}
\usage{
parse_rhessys_cmdout(cmdout)
}
\arguments{
\item{cmdout}{The command output string from RHESSysIOinR that would be input for running RHESSys.}
}
\value{
A list containing the program and its arguments.
}
\description{
Parse RHESSys command output
}
