% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfile_utils.R
\name{patches2patchfamily}
\alias{patches2patchfamily}
\title{Aggregate patch data to patch families}
\usage{
patches2patchfamily(output_df)
}
\arguments{
\item{output_df}{Data table with patch-level outputs}
}
\description{
Aggregate patch data to patch families
}
