% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{pfam_extract_world}
\alias{pfam_extract_world}
\title{Extract a patch family sub-world}
\usage{
pfam_extract_world(familyID, world)
}
\arguments{
\item{familyID}{Patch family identifier}

\item{world}{Worldfile data frame with family information}
}
\description{
Extract a patch family sub-world
}
