% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpdf_allvars.R
\name{plotpdf_allvars}
\alias{plotpdf_allvars}
\title{plotpdf_allvars}
\usage{
plotpdf_allvars(
  out_dir,
  out_name = "plots",
  out_path = "plots",
  pattern = "_basin",
  step = "monthly",
  pdfwidth = 7,
  pdfheight = 7,
  hide_legend = F,
  summary_plots = F,
  run_limit = 25,
  auto_vars = T,
  runs = NULL
)
}
\arguments{
\item{out_dir}{Output directory to search for rhessys output files}

\item{out_name}{Base name of output pdf. Unique timedate is added to this string. If left null 'plots' is used.}

\item{out_path}{Path to folder where pdf will be written. If doesn't exist, will be created.}

\item{pattern}{Text pattern to use when searching for output files. Default is '_basin'.}

\item{step}{Time step for plots. Only options are 'monthly' or 'yearly'.}

\item{pdfwidth}{Width for output pdf in inches.}

\item{pdfheight}{Height for output pdf in inches.}

\item{hide_legend}{TRUE/FALSE if legend should be hidden in output plot.}

\item{summary_plots}{TRUE/FALSE if data should be summarised instead of individual runs.}

\item{run_limit}{Number of runs which automatically trigger summary statistics instead of individual run plots.}

\item{auto_vars}{TRUE/FALSE to automatically generate relevent variables: evapotranspiration}

\item{runs}{Optional vector of run IDs, to subset the total dataset. Defaults to NULL. May not work}
}
\description{
Reads, aggregates, and plots all output variables to a pdf. Assumes output using output filters. Optionally can aggregate, create summary statistics, hide legend. Uses a pattern input to select one or many csv outputs in an output directory. Currently designed to work with basin scale output.
}
