% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utils.R
\name{polaris2texture}
\alias{polaris2texture}
\title{Derive USDA soil texture classes from sand/clay rasters}
\usage{
polaris2texture(basin, sand, clay, plot_out, writeplots = TRUE, plots = TRUE)
}
\arguments{
\item{basin}{SpatRaster or path to basin mask raster}

\item{sand}{SpatRaster or path to sand percentage raster}

\item{clay}{SpatRaster or path to clay percentage raster}

\item{plot_out}{File path for optional PDF plot}

\item{writeplots}{Logical; write PDF output}

\item{plots}{Logical; show and/or generate plots}
}
\value{
SpatRaster of USDA soil texture class IDs
}
\description{
Derive USDA soil texture classes from sand/clay rasters
}
