% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{read_pars_table}
\alias{read_pars_table}
\title{Read a parameter table}
\usage{
read_pars_table(
  out_dir = NULL,
  csv = NULL,
  listformat = F,
  table_name = "all_def_changes"
)
}
\arguments{
\item{out_dir}{Output directory containing params subfolder}

\item{csv}{Optional path to a parameter CSV}

\item{listformat}{Logical; return list instead of data frame}

\item{table_name}{Filename pattern for parameter table}
}
\description{
Read a parameter table
}
