% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{rhessysIO2pronghorn}
\alias{rhessysIO2pronghorn}
\title{Transfer RHESSys inputs and commands to pronghorn HPC}
\usage{
rhessysIO2pronghorn(
  rhout,
  name,
  rh_bin_replace = "/RHESSys/rhessys7.5",
  dest,
  usr,
  input_rhessys = NULL,
  input_hdr = NULL,
  input_def_pars = NULL,
  input_tec_data = NULL,
  output_filter = NULL,
  filelist = NULL,
  transfer_method = "scp",
  rsa_loc = "~/rsa_key"
)
}
\arguments{
\item{rhout}{Character vector of RHESSys commands}

\item{name}{Label for generated files}

\item{rh_bin_replace}{Replacement RHESSys binary path on HPC}

\item{dest}{Remote destination path}

\item{usr}{Remote user (with hostname)}

\item{input_rhessys}{RHESSys inputs list}

\item{input_hdr}{Header list of definition files}

\item{input_def_pars}{Definition parameter list}

\item{input_tec_data}{Tec events data frame}

\item{output_filter}{Output filter list}

\item{filelist}{Optional precomputed file list}

\item{transfer_method}{Transfer method (`"scp"` or `"rsync"`)}

\item{rsa_loc}{Path to SSH key}
}
\description{
Transfer RHESSys inputs and commands to pronghorn HPC
}
