% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{rhout_write_for_hpc}
\alias{rhout_write_for_hpc}
\title{Format RHESSys run commands for HPC submission}
\usage{
rhout_write_for_hpc(rhout, name, rh_bin_replace = NULL)
}
\arguments{
\item{rhout}{Character vector of RHESSys commands}

\item{name}{Label to include in the output filename}

\item{rh_bin_replace}{Optional replacement RHESSys binary path}
}
\description{
Format RHESSys run commands for HPC submission
}
