% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{runif_sample}
\alias{runif_sample}
\title{Sample uniformly within a range}
\usage{
runif_sample(X, n)
}
\arguments{
\item{X}{Numeric vector of length 2 (min, max)}

\item{n}{Number of samples}
}
\description{
Sample uniformly within a range
}
