% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{select_familyID_per_vegID_MSR}
\alias{select_familyID_per_vegID_MSR}
\title{Select patch family closest to median elevation and cover for a rule ID}
\usage{
select_familyID_per_vegID_MSR(x, world)
}
\arguments{
\item{x}{Rule identifier}

\item{world}{Worldfile data frame with rule information}
}
\description{
Select patch family closest to median elevation and cover for a rule ID
}
