% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{select_mean_vegID_patch}
\alias{select_mean_vegID_patch}
\title{Select patch closest to median elevation and cover for a veg ID}
\usage{
select_mean_vegID_patch(x, world)
}
\arguments{
\item{x}{Vegetation parameter ID}

\item{world}{Worldfile data frame}
}
\description{
Select patch closest to median elevation and cover for a veg ID
}
