% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{select_pfam}
\alias{select_pfam}
\title{Select representative patch family by rule}
\usage{
select_pfam(rule, world)
}
\arguments{
\item{rule}{Rule identifier}

\item{world}{Worldfile data frame with rule information}
}
\description{
Select representative patch family by rule
}
