% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_rhessys_folders}
\alias{setup_rhessys_folders}
\title{Create a standard RHESSys project folder layout}
\usage{
setup_rhessys_folders(base_dir)
}
\arguments{
\item{base_dir}{Base directory in which to create folders}
}
\description{
Create a standard RHESSys project folder layout
}
