% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utils.R
\name{soil_texture_plot}
\alias{soil_texture_plot}
\title{Plot USDA soil texture map and table summary}
\usage{
soil_texture_plot(soil_texture, plot_out, writeplots = T)
}
\arguments{
\item{soil_texture}{SpatRaster of soil texture class IDs}

\item{plot_out}{File path to write PDF plot}

\item{writeplots}{Logical; write PDF output}
}
\description{
Plot USDA soil texture map and table summary
}
