% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tec_copy_redef}
\alias{tec_copy_redef}
\title{Copy redefine TEClist entry to a new date}
\usage{
tec_copy_redef(input_redef, redef_date, worldfile, overwrite = F)
}
\arguments{
\item{input_redef}{Path to input redefine tec file}

\item{redef_date}{Date to copy}

\item{worldfile}{Path to worldfile}

\item{overwrite}{Logical; overwrite output}
}
\description{
Copy redefine TEClist entry to a new date
}
