% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{unif_sample_all}
\alias{unif_sample_all}
\title{Sample uniformly for all parameters}
\usage{
unif_sample_all(par_ranges, n_pars_each)
}
\arguments{
\item{par_ranges}{List of parameter ranges}

\item{n_pars_each}{Number of samples per parameter}
}
\description{
Sample uniformly for all parameters
}
