% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{watbal_basin_of_multi}
\alias{watbal_basin_of_multi}
\title{Compute water balance for multiple basin outputs}
\usage{
watbal_basin_of_multi(out_dir)
}
\arguments{
\item{out_dir}{Directory containing basin output CSV files}
}
\description{
Compute water balance for multiple basin outputs
}
