% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utils.R
\name{wbox_dem2streams_gauge_basin}
\alias{wbox_dem2streams_gauge_basin}
\title{Derive streams, snapped gauge, and basin extent}
\usage{
wbox_dem2streams_gauge_basin(
  source_dem,
  source_gauge,
  res,
  stream_threshold = 100,
  gauge_snap_dist = 90,
  output_dir,
  tmp_dir = file.path(output_dir, "wb_tmp"),
  plots = T,
  writeplots = T,
  overwrite = T
)
}
\arguments{
\item{source_dem}{Path to source DEM raster}

\item{source_gauge}{Path to gauge point}

\item{res}{Target resolution (map units)}

\item{stream_threshold}{Flow accumulation threshold for streams}

\item{gauge_snap_dist}{Snap distance for gauge (map units)}

\item{output_dir}{Output directory}

\item{tmp_dir}{Temporary working directory}

\item{plots}{Logical; show plots}

\item{writeplots}{Logical; write plots to PDF}

\item{overwrite}{Logical; overwrite outputs}
}
\description{
Derive streams, snapped gauge, and basin extent
}
