% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utils.R
\name{wbox_slope_aspect_horizons}
\alias{wbox_slope_aspect_horizons}
\title{Compute slope, aspect, and horizons}
\usage{
wbox_slope_aspect_horizons(
  dem_brch,
  output_dir,
  tmp_dir = file.path(output_dir, "wb_tmp"),
  plots = T,
  writeplots = T,
  overwrite = T
)
}
\arguments{
\item{dem_brch}{Path to breached DEM raster}

\item{output_dir}{Output directory}

\item{tmp_dir}{Temporary working directory}

\item{plots}{Logical; show plots}

\item{writeplots}{Logical; write plots to PDF}

\item{overwrite}{Logical; overwrite outputs}
}
\description{
Compute slope, aspect, and horizons
}
