% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utils.R
\name{wbox_subbasins}
\alias{wbox_subbasins}
\title{Delineate subbasins using Whitebox}
\usage{
wbox_subbasins(
  dem_brch,
  streams,
  output_dir,
  stream_threshold,
  tmp_dir = file.path(output_dir, "wb_tmp"),
  plots = T,
  writeplots = T,
  overwrite = T
)
}
\arguments{
\item{dem_brch}{Path to breached DEM raster}

\item{streams}{Path to streams raster}

\item{output_dir}{Output directory}

\item{stream_threshold}{Flow accumulation threshold for streams}

\item{tmp_dir}{Temporary working directory}

\item{plots}{Logical; show plots}

\item{writeplots}{Logical; write plots to PDF}

\item{overwrite}{Logical; overwrite outputs}
}
\description{
Delineate subbasins using Whitebox
}
