% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utils.R
\name{wbox_subbasins_vis}
\alias{wbox_subbasins_vis}
\title{Visualize subbasins with leaflet}
\usage{
wbox_subbasins_vis(subbasins, streams, min_subbasin_size = 20)
}
\arguments{
\item{subbasins}{Path to subbasins raster}

\item{streams}{Path to streams raster}

\item{min_subbasin_size}{Minimum subbasin size (ha)}
}
\description{
Visualize subbasins with leaflet
}
