% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitebox_utils.R
\name{wbox_trim_to_basin_overwrite}
\alias{wbox_trim_to_basin_overwrite}
\title{Trim maps to basin extent}
\usage{
wbox_trim_to_basin_overwrite(
  maps,
  basin,
  output_dir,
  plots = T,
  writeplots = T
)
}
\arguments{
\item{maps}{Named list of map paths (includes basin, streams, DEM, optional gauge_snap)}

\item{basin}{Path to basin raster}

\item{output_dir}{Output directory}

\item{plots}{Logical; show plots}

\item{writeplots}{Logical; write plots to PDF}
}
\description{
Trim maps to basin extent
}
