% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{win_banner}
\alias{win_banner}
\title{Display a Windows toast notification}
\usage{
win_banner(txt, src = "R - Positron")
}
\arguments{
\item{txt}{Message to display in the toast notification}

\item{src}{Source application name for the toast notification}
}
\description{
Display a Windows toast notification
}
