% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{world_add_familyID_RuleID}
\alias{world_add_familyID_RuleID}
\title{Add family and rule identifiers to a worldfile table}
\usage{
world_add_familyID_RuleID(world)
}
\arguments{
\item{world}{Worldfile data frame containing `family_ID` entries}
}
\description{
Add family and rule identifiers to a worldfile table
}
