% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{world_add_level_i}
\alias{world_add_level_i}
\title{Add numeric level indices to a worldfile table}
\usage{
world_add_level_i(world)
}
\arguments{
\item{world}{Worldfile data frame}
}
\description{
Add numeric level indices to a worldfile table
}
