% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinup_utils.R
\name{world_add_patch_vegparmIDcol}
\alias{world_add_patch_vegparmIDcol}
\title{Add patch-level vegetation parameter identifiers}
\usage{
world_add_patch_vegparmIDcol(world)
}
\arguments{
\item{world}{Worldfile data frame}
}
\description{
Add patch-level vegetation parameter identifiers
}
