% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfile_utils.R
\name{world_add_var}
\alias{world_add_var}
\title{Insert a variable into a worldfile table}
\usage{
world_add_var(world, var, value, loc_var)
}
\arguments{
\item{world}{Worldfile data frame}

\item{var}{Variable name to insert}

\item{value}{Value to assign to the new variable}

\item{loc_var}{Existing variable name after which to insert}
}
\value{
Updated worldfile data frame with the new row inserted
}
\description{
Insert a variable into a worldfile table
}
