% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{worldstate}
\alias{worldstate}
\title{Locate a worldfile .state companion file}
\usage{
worldstate(worldfile)
}
\arguments{
\item{worldfile}{Path to worldfile}
}
\description{
Locate a worldfile .state companion file
}
