% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_basestation.R
\name{write_basestation}
\alias{write_basestation}
\title{write_basestation}
\usage{
write_basestation(
  dest,
  year_start_index = 1900,
  search_dist,
  netcdf_var_x = "lon",
  netcdf_var_y = "lat",
  precip_multiplier = 0.001,
  rhum_multiplier = 0.01,
  temp_unit = "K",
  IDs,
  dem,
  lat,
  lon,
  x = NULL,
  y = NULL,
  lai = 3,
  screen_height = 10,
  netcdf_tmax_filename,
  netcdf_var_tmax = "daily_maximum_temperature",
  netcdf_tmin_filename,
  netcdf_var_tmin = "daily_minimum_temperature",
  netcdf_rain_filename,
  netcdf_var_rain = "precipitation_amount",
  netcdf_huss_filename = NULL,
  netcdf_var_huss = "daily_mean_specific_humidity",
  netcdf_rmax_filename = NULL,
  netcdf_var_rmax = "daily_maximum_relative_humidity",
  netcdf_rmin_filename = NULL,
  netcdf_var_rmin = "daily_minimum_relative_humidity",
  netcdf_rsds_filename = NULL,
  netcdf_var_rsds = "daily_mean_shortwave_radiation_at_surface",
  netcdf_was_filename = NULL,
  netcdf_var_was = "daily_mean_wind_speed"
)
}
\description{
write_basestation
}
