% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{write_param_table}
\alias{write_param_table}
\title{Write parameter table}
\usage{
write_param_table(input_def_pars, outfile_basename = "all_def_changes")
}
\arguments{
\item{input_def_pars}{Definition parameter list}

\item{outfile_basename}{Base name for output CSV}
}
\description{
Write parameter table
}
