% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_utils.R
\name{write_rhessys_launch_json_from_cmd}
\alias{write_rhessys_launch_json_from_cmd}
\title{Write RHESSys launch JSON from command output}
\usage{
write_rhessys_launch_json_from_cmd(
  cmdout,
  launch_path = ".vscode/launch.json",
  config_name = "Debug RHESSys",
  cwd = "${workspaceFolder}",
  miDebuggerPath = "/usr/bin/gdb",
  program_override = NULL
)
}
\arguments{
\item{cmdout}{The command output string from RHESSys.}

\item{launch_path}{The path to the launch JSON file.}

\item{config_name}{The name of the debug configuration.}

\item{cwd}{The current working directory for the debugger.}

\item{miDebuggerPath}{The path to the MI debugger.}

\item{program_override}{Optional program path override.}
}
\value{
A list containing the path to the launch JSON and the launch object.
}
\description{
Write RHESSys launch JSON from command output
}
