% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPC_utils.R
\name{write_rules_file}
\alias{write_rules_file}
\title{Write RHESSys rules file from LPC-derived rules}
\usage{
write_rules_file(
  rules,
  output_rules_file,
  vegid_tree = 1,
  vegid_shrub = 5,
  vegid_herb = 3,
  vegid_other = 4,
  thin_vegid_mod = 0,
  strata_ct = NULL
)
}
\arguments{
\item{rules}{Data frame of rules and cover fractions}

\item{output_rules_file}{Path to write the rules file}

\item{vegid_tree}{Definition file ID for trees}

\item{vegid_shrub}{Definition file ID for shrubs}

\item{vegid_herb}{Definition file ID for herbs}

\item{vegid_other}{Definition file ID for other vegetation}

\item{thin_vegid_mod}{Offset to apply for thinned vegetation IDs}

\item{strata_ct}{Optional vector of strata counts per rule}
}
\description{
Write RHESSys rules file from LPC-derived rules
}
