% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defpars_utils.R
\name{write_updated_def_files}
\alias{write_updated_def_files}
\title{Write updated definition files}
\usage{
write_updated_def_files(input_def_pars, input_hdr, filename_ext = NULL)
}
\arguments{
\item{input_def_pars}{Definition parameter list}

\item{input_hdr}{Header lines to prepend}

\item{filename_ext}{Optional filename extension}
}
\description{
Write updated definition files
}
