% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_world}
\alias{write_world}
\title{Write a worldfile table back to disk}
\usage{
write_world(world, path)
}
\arguments{
\item{world}{Data frame produced by `read_world`}

\item{path}{Output file path for the worldfile}
}
\description{
Write a worldfile table back to disk
}
