% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_eval_generic.R
\name{cal_eval_generic}
\alias{cal_eval_generic}
\title{cal_eval_generic}
\usage{
cal_eval_generic(
  sim,
  obs,
  compare_col,
  eval_monthly = F,
  obs_is_monthly = F,
  echo_best = T,
  agg_by = "mean"
)
}
\description{
Calculates calibration statistics: nse, lognse, pbias, rmse, and r2, at either daily or monthly.
}
