% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_eval_generic.R
\name{cal_eval_generic}
\alias{cal_eval_generic}
\title{cal_eval_generic}
\usage{
cal_eval_generic(sim, obs, col, monthly = F, echo_best = T)
}
\description{
Takes either the paths to simulated and observed GENERIC DATA he dataframes/datatables themselves, and calculates
nse, lognse, pbias, rmse, and r2, at either daily or monthly. Requires get_basin_daily from rhutils and the datatable packges.
}
