% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_utils.R
\name{check_template}
\alias{check_template}
\title{Summarize RHESSys template contents}
\usage{
check_template(template)
}
\arguments{
\item{template}{Path to RHESSys template file}
}
\description{
Summarize RHESSys template contents
}
