% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_output.R
\name{collect_output}
\alias{collect_output}
\title{collect_output}
\usage{
collect_output(source_dir = "./", basename = "rh_out_", output_dir = "output")
}
\description{
Collects the output files data and parameter files from a rhessys run and auto-generates a folder to place them in
}
