% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_alert}
\alias{sim_alert}
\title{Play a visual and audio alert when simulations complete}
\usage{
sim_alert(txt = "Simulations Complete", fg = 46, bg = 24)
}
\arguments{
\item{txt}{Message to display}

\item{fg}{Foreground ANSI color code}

\item{bg}{Background ANSI color code}
}
\description{
Play a visual and audio alert when simulations complete
}
